Implemente CycleContext to activity page + refactor code to dynamically change the list based on cycles
