+++ src/pages/Activity/index.tsx
@@ -1,9 +1,15 @@
+import { useContext } from 'react'
+import { formatDistanceToNow } from 'date-fns'
 import { ActivityContainer, ActivityList, Status } from './styles'
+import { CyclesContext } from '../../context/CyclesContext'
 
 export function Activity() {
+  const { cycles } = useContext(CyclesContext)
+
   return (
     <ActivityContainer>
       <h1>Activity</h1>
+
       <ActivityList>
         <table>
           <thead>
@@ -15,30 +21,30 @@ export function Activity() {
             </tr>
           </thead>
           <tbody>
-            <tr>
-              <td>Task</td>
-              <td>20 min</td>
-              <td>2 months ago</td>
-              <td>
-                <Status statusColor="red">Interrupted</Status>
-              </td>
-            </tr>
-            <tr>
-              <td>Task</td>
-              <td>20 min</td>
-              <td>2 months ago</td>
-              <td>
-                <Status statusColor="yellow">Ongoing</Status>
-              </td>
-            </tr>
-            <tr>
-              <td>Task</td>
-              <td>20 min</td>
-              <td>2 months ago</td>
-              <td>
-                <Status statusColor="green">Finished</Status>
-              </td>
-            </tr>
+            {cycles.map((cycle) => {
+              return (
+                <tr key={cycle.id}>
+                  <td>{cycle.task}</td>
+                  <td>{cycle.minutesAmount} minutes</td>
+                  <td>
+                    {formatDistanceToNow(cycle.startDate, { addSuffix: true })}
+                  </td>
+                  <td>
+                    {cycle.finishedDate && (
+                      <Status statusColor="green">Completed</Status>
+                    )}
+
+                    {cycle.interruptedDate && (
+                      <Status statusColor="red">Interrupted</Status>
+                    )}
+
+                    {!cycle.finishedDate && !cycle.interruptedDate && (
+                      <Status statusColor="yellow">In Progress</Status>
+                    )}
+                  </td>
+                </tr>
+              )
+            })}
           </tbody>
         </table>
       </ActivityList>
